Refactored a little bit
