+++ traffic_ops/traffic_ops_golang/api/api_test.go
@@ -1,0 +1,14 @@
+package api
+
+import "testing"
+
+func TestCamelCase(t *testing.T) {
+
+	testStrings := []string{"hello_world", "trailing_underscore_", "w_h_a_t____"}
+	expected := []string{"helloWorld", "trailingUnderscore", "wHAT"}
+	for i, str := range testStrings {
+		if toCamelCase(str) != expected[i] {
+			t.Errorf("expected: %v error, actual: %v", expected[i], toCamelCase(str))
+		}
+	}
+}
