+++ traffic_ops/traffic_ops_golang/dbhelpers/db_helpers_test.go
@@ -71,14 +71,3 @@ FROM table t
 	}
 
 }
-
-func TestCamelCase(t *testing.T) {
-
-	testStrings := []string{"hello_world", "trailing_underscore_", "w_h_a_t____"}
-	expected := []string{"helloWorld", "trailingUnderscore", "wHAT"}
-	for i, str := range testStrings {
-		if toCamelCase(str) != expected[i] {
-			t.Errorf("expected: %v error, actual: %v", expected[i], toCamelCase(str))
-		}
-	}
-}
